modified:   api/Controllers/RefreshTokensController.cs 	modified:   api/Controllers/UsersController.cs 	modified:   api/ExtensionMethods/JwtExtensionMethods.cs 	new file:   api/ExtensionMethods/PasswordExtensionMethods.cs
