+++ api/ExtensionMethods/JwtExtensionMethods.cs
@@ -1,11 +1,13 @@
+using api.Models;
 using Microsoft.IdentityModel.Tokens;
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
+using System.Security.Cryptography;
 using System.Text;
 
 namespace api.ExtensionMethods;
 
-public static class StringExtensions
+public static class JwtExtensions
 {
     public static JwtSecurityToken GenerateJwt(this string userId, string jwtKey)
     {
@@ -21,5 +23,22 @@ public static class StringExtensions
         );
 
         return token;
+    }
+
+    public static RefreshToken GenerateRefreshToken(this string userId)
+    {
+        var randomNumber = new byte[64];
+        using var rng = RandomNumberGenerator.Create();
+        rng.GetBytes(randomNumber);
+
+        string token = Convert.ToBase64String(randomNumber);
+
+        var refreshToken = new RefreshToken{
+            UserId = userId,
+            Token = token,
+            ExpiresAt = DateTime.UtcNow.AddHours(24)
+        };
+
+        return refreshToken;
     }
 }
